refactor(data-tracks): move the Room lifecycle hooks to their own file
Room+DataTrack.swift had been sitting on the 400-line limit, with the
last few changes paid for by trimming comments. setupDataTracks and
cleanUpDataTracks are Room's side of the subsystem rather than the
coordinator's, so they read fine apart from it.
